pyproject.toml
[project]
name = "mqtt-client"
version = "0.1.0"
description = ""
authors = [
  { name = "Chew Chee Wai", email = "chew.cheewai@ynytechnology.com.my" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "paho-mqtt>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
]

[tool.pyright]
include = ["."]
venv = ".venv"
venvPath = ["/workspace/rnd/asusiot/mqtt-client/.venv"]
